+++ packages/terra-core-docs/src/terra-dev-site/test/paginator/ControlledProgressivePaginator.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Button from 'terra-button';
 import classNames from 'classnames/bind';
-import ControlledProgressivePaginator from '../../../ControlledProgressivePaginator';
+import ControlledProgressivePaginator from 'terra-paginator/lib/ControlledProgressivePaginator';
 import styles from './ControlledPaginatorTestCommon.module.scss';
 
 const cx = classNames.bind(styles);