+++ packages/terra-core-docs/src/terra-dev-site/test/paginator/PaginatorWithoutItemCount.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Paginator from '../../../Paginator';
+import Paginator from 'terra-paginator';
 
 const PaginatorWithoutItemCountExample = () => (
   // eslint-disable-next-line no-console